Solution to Input and Print an Integer

+++ Input and Print an Integer@@6788d5012db25efec946f939/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    float num;
+    int num;
     scanf("%d",&num);
     printf("You entered: %.2d",num);
     return 0;
